[ext4] Fix unsafe construct
We can't just `Arc::from_raw` on a `Box::into_raw` as this will expect
the pointer to be part of an `ArcInner` and as such improperly use heap
memory located prior to the pointer as part of this struture, notably
the Scudo chunk header.

This was noticed when trying to update Scudo (fxr/343518) and lead to
various errors.

Change-Id: Ib323cf6c5dd39c8ed27a3ac36b7a0a0ff7f04902
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427095
Reviewed-by: Victor Chang <vfcc@google.com>
Testability-Review: Benjamin Wright <benwright@google.com>
Commit-Queue: Kostya Kortchinsky <kostyak@google.com>