[GN SDK] Adding fconfig properties for emulator
This adds `emu-bucket` and `emu-image` to the list of
valid properties for `fconfig.sh`.

These are used as defaults for the emulator. They are different
than the default properties for fpave.sh and fserve.sh because it is
a common use case to switch between a device and an emulator.

Bug: b/168540844

Change-Id: Iaae8b99332b0e4a9b1b1352eb6701ebb7cf3f753
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427496
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jonathan Koren <jdkoren@google.com>
Testability-Review: Jonathan Koren <jdkoren@google.com>