[testrunner] Remove support for custom commands and runnerless tests
Recent changes to the build system have made it impossible for a test
spec to contain a command field. The changes in this CL are a nice
simplification.

Also remove support for directly invoking a path on Fuchsia. This
support is unused, and we should not allow it, as it would bypass
per-test timeout enforcement.

Fixed: 58561
Change-Id: Ibacb4d9aae442c83658e30dfd1a5c1ff5d37983a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426557
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>