[zbitl] Don't assume test payloads are NUL-free
This is a brittle assumption, which will be invalidated once we start
looking at expected ranges of storage across multiple items. In this
change, we alias std::string so as not to convey that we are dealing in
text - and replace all usage of EXPECT_STR_EQ with EXPECT_BYTES_EQ.

Also, we sneak in a standardization to make use of `storage_type` and
`payload_type` in the test trait signatures.

Bug: 57764
Change-Id: Ie3dec71842e7722f43daacb7f339c234b2424474
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427355
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>