[wlan][policy] matching mixed networks
This change allows us to auto connect to mixed networks if the network
is saved as either of the security types in the mixed security type. Ie
if we save a WPA3 (or WPA2) network and see it as a WPA2/WPA3 network we
will auto connect to it.

Multiply: wlancfg-tests
Bug: 53834
Test: fx test wlancfg-tests
Change-Id: Ibe4821572faabd299d2c8abf8108528d29e7d1b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/424162
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Testability-Review: Marc Khouri <mnck@google.com>