[usb_bulk] Fix MacOs interface.
This fixes two bugs observable in test cases of usb_bulk_lib_test.

Currently, ClearPipeStallBothEnds is called on each endpoint when the
interface is opened. However, when this logic is present, the test
case test_zedmon_read_parameter fails with a timeout every time but
the first that it runs for a given boot of the Zedmon firmware. (This
failure is representative of errors encountered when attempting to
utilize the usb_bulk library elsewhere.) The reason for the failure is
unclear, but neither is the reason for clearing pipe stalls.
Furthermore the stall-clearing dates to 2010 or earlier; MacOs
behavior may well have changed in the interim.

Additionally, the USB interface is explicitly closed and released.
Prior to that modification, test_zedmon_enumeration is flaky. The
USBInterfaceClose and Release calls are illustrated in Apple
documentation:
https://developer.apple.com/library/archive/documentation/DeviceDrivers/Conceptual/USBBook/USBDeviceInterfaces/USBDevInterfaces.html#//apple_ref/doc/uid/TP40002645-TPXREF101

Test (manual, requires Zedmon hardware): fx test usb_bulk_lib_test

Fixed: 58426
Change-Id: I9ebf4082877f4bcd83dcdce44a8866282b03a098
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418850
Reviewed-by: Matthew Boetger <boetger@google.com>
Commit-Queue: Onath Claridge <claridge@google.com>
Testability-Review: Matthew Boetger <boetger@google.com>