[audio_core] Device clock is clone of driver clock
In preparation of synthesizing the hardware clock from position
notifications, AudioDriver maintains a separate AudioClock that it
will rate-adjust as part of the clock recovery process. The
existing AudioClock (passed upward to the output or input mix) is
now a read-only clone of the AudioDriver's underlying zx::clock.

We only create this separate clock, and perform hardware clock
recovery, if the device is NOT in the Monotonic clock domain.
For this reason, AudioDriver must wait until it has retrieved the
clock domain from the actual audio driver, before creating the
AudioClock.

Bug:  46631
Test: AudioDriver/AudioDevice tests reworked to accommodate

Change-Id: I7a54c624ad96330b01d819e0e8fa0760777ec277
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427151
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tom Bergan <tombergan@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>