[testrunner] Reconnect if failed to run snapshot.
We only try to reconnect to SSH in between tests, so if we lose the
connection after the last test and before we run the snapshot, it will
fail with "ssh is diconnected". We should try to reconnect and retry
running the snapshot as well.

Bug: 60039
Change-Id: Ia26acd0ba77d02aaaafe29393b7ceb5b23e525f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427267
Reviewed-by: Oliver Newman <olivernewman@google.com>
Testability-Review: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>