[build] Fix LTO build.
A previous CL moved the build of the Gigaboot bootloader to the
Fuchsia build. This binary in a Win32 PE+ file that needs to be
linked with the "lld-link" linker, which expects slightly different
command-line parameter names to control lto parallelism and caching.

This broke the LTO build, which is not on CQ, hence why it went
unnotticed at the time of submit. This CL fixes the issue by ensuring
the right parameters are used for this specific case.

Fixed: 60060
Testing: Fix the LTO build (e.g. --variant=thintlo)
Change-Id: I519f6c637a4e1f5cf330884ac954741fcca94425
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427220
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>