[usb] Fix invalid assumption about port ordering
Per USB 2.0 specification revision 2.0 section 9.1.2,
enumeration order on a hub is not guaranteed to be sequential
in order of port number. This was accurately reflected in my hub simulation,
but the test didn't account for this behavior. This CL updates the enumeration
test to allow for enumeration to occur in any order.

Test: fx test "//src/devices/usb/drivers/usb-hub-rewrite:usb-hub-package_test_usb-hub-package_component(//build/toolchain/fuchsia:x64)"
Change-Id: I15882b01f2806798f7b8fc336221deaa1c6ba127
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427463
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>