[build] Add documentation for gen-latest-commit-date args.
fxb/58677 is pointing to the fact that the formats for our backstop
files are not well documented and will be even less well documented once
we remove the kernel clock and its command line argument.

This CL adds a little more documentation by describing the expected
output in the script that generates the files.

Bug: 58677
Change-Id: I4d2bdafc25d6192cbf3de0971484df87a6b8d24d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426015
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>