[build] Set *_SYMBOLIZER_PATH in environment of host_tool_action() runs
The Zircon build always had machinery intended for this, but it
hadn't been fully wired up.  Note that the Zircon build already
runs the unstripped binaries for host_tool_action() though it
heretofore failed to point to the symbolizer so it would only get
full symbolization if llvm-symbolizer or addr2line that works is
in the user's PATH already.  The Fuchsia build already takes care
to set the environment variables via gn_run_binary.sh, but it
runs the stripped binaries so available symbolization is minimal.

Change-Id: Id661ec7f112a6b1f55a3bb5d111ce6b1ff5cef23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426564
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>