[build] Allow metadata for shared_library() instances.
Building the Musl C library with the Fuchsia build requires passing
metadata to a shared_library() instance. This was not possible, and
this CL fixes the issue.

For the record, the metadata is used to populate the sysroot.

Bug: 54161
Testing: Should not change the build.

Change-Id: I3b13882ab1630e267d80b33c05c385e124040836
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427856
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>