[goldfish] Support VkBuffer allocation on goldfish HostVisibleHeap.
Now goldfish HostVisibleHeap can allocate VMOs and set up
renderControl Buffers for VkBuffer usage.

Currently whether it is a Buffer or it's an Image is determined by
the constraints: If the constraints contains ImageFormatConstraints,
then we regard it as a Buffer, otherwise we regard it as an Image.

TEST=add tests testing buffer allocation to goldfish_test
Bug: 59805

Change-Id: Ibfe5ba2325bdec521f40a50a995c83e3bf31aa64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426320
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Testability-Review: David Reveman <reveman@google.com>