[weave] Reduce the max sleep time for select to 1 second
Currently the max sleep time for select is set to 10 seconds.
This causes the runloop to sleep longer and timer/task processing
is delayed. Reduce it to 1 second. And also wakeup the
runloop after posting a task, so that the task runs immediately.

Bug: 59242
Test: Manually verified that the runloop wakes up within 1 second
      and whenever a task is posted.
Test: fx test weavestack-unittests.

Change-Id: Ia03d26a5c2b3fdd0f250e0a51c9d2960b13e6525
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423038
Commit-Queue: Sundarajan Srinivasan <sundarsrini@google.com>
Testability-Review: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>