[codesize] Simple test for fx codesize
This tests running "fx codesize --help" exits successfully
with code 0.

Currently, this cannot run on CQ because a dart binary
isn't available in that environment, but plans for
support for the runtime in CQ are in progress and
it can be run with fx self-test in the meantime.

Test: fx self-test fx_codesize_test

Change-Id: I719d59edb59f13280106b810aeeaa086a8cae340
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427017
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>