[tools][fidlc] Improve parsing of response files.
This CL removes an important restriction from fidlc, i.e. that
a response file, if used on the command-line, is the only argument
allowed when invoking the program.

Instead, the response file is just read and inserted into the
current command-line, allowing the caller to provide other arguments
before or after the response file, or even using several ones in
a single call. Response files still do not nest though.

This CL is necessary to simplify build unification. For more
context, see the comments on this Gerrit CL:

  https://fuchsia-review.googlesource.com/c/fuchsia/+/427858

Bug: 54161
Testing: Should not change the build.
Change-Id: I7eb0072be6333383b9b3e3176100aa4a9b9643e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428074
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Testability-Review: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>