[SetUI] Introduce Event to Handler MessageHub
This changelist introduces the Event enum to the Handler
MessageHub payload and defines the Changed event
type there. For this event, the SettingType has been
dropped as it is redundant for the SettingProxy.

This change is in anticipation for an upcoming addition of
new event types.

Bug: 58181
Test: fx test -o setui_service_tests
Change-Id: Ib1025ed5dd0570f7aca80a8b7fc3147ba49e912f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426314
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: William Xiao <wxyz@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>