[build] Ignore variant_target in toolchain_util_action()
This fixes the definition of toolchain_util_action() when
it is called from the Fuchsia build. When migrating the
arm64 boot shims, the `variant_target` parameter will be
set.

Instead of asserting, just ignore the parameter because
it is not used in practice: i.e. this only makes sense
when the kernel is built with a different variant than
the boot shims, something that is not supported anymore.

Bug: 59918
Testing: Should not change the build.

Change-Id: Ib624257eea48320527f96515219be6861c0e6dd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428538
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>