[build][alln] Prepare zircon script for host tool migration.
Host tools like lz4 and zbi are soon going to be moved entirely
to the Fuchia build. This CL modifies a Zircon script that is
invoked from the Fuchsia build, but expects some files to
be located in the Zircon output directory.

This simply replaces all occurences of BUILD_DIR with
ZIRCON_BUILD_DIR, and adds a new variable ROOT_BUILD_DIR
to point to the Fuchsia output directory instead.

The latter can be set manually with -R <build-path>, or
is simply guessed from Zircon build dir otherwise.

As more items are moved from the Zircon to the Fuchsia builds,
lines of this script will be updated progressively.

+ Remove obsolete source file that is no longer compiled
  by anything, and the script lines that reference its binary.

Bug: 49742
Testing: Should not change the build.
Change-Id: I9dd2fef6b6cbf44c249b7e32903180a97bd8f295
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428536
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>