[build][alln] Restore obsolete legacy libraries entries in .gitignore
This restores a few entries in zircon/public/lib/.gitignore for a few
libraries that were already migrated to the Fuchsia build. The reason
to do this is to avoid lingering untracked directories to appear
in 'git status', because the corresponding files are still in the
workspace (they were not managed by git but auto-generated).

It doesn't change the build in any way, it just makes it easier
to avoid polluting workspaces with untracked files during build
unification.

A comment is also added to the file explaining why this is needed
and how to manage this file during future legacy target migrations.

Bug: 58463
Testing: No change to the build.
Change-Id: Ic6ccb173b10db1214f6d8c896ca6f090a246a8e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428535
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>