[event-queue] Add try_flush() API
This change adds a try_flush method to event-queue's control handle that
produces a future that completes when all currently queued events are
dispatched and acknowledged by actively connected clients.

Normally, users of event-queue want to fire off events to any listeners
and then forget about them, but if a service emitting events wants to
stop or restart, for example, it needs a mechanism to wait for all
enqueued events to be sent to listeners before it can do so.

Note that, normally, it is by-design that listeners of events can't stop
the service from making forward progress.  To ensure that a misbehaving
listener can't block a service from stopping, this API will timeout
after a configurable duration.

Fixed: 60146
Change-Id: Ic99c13e3c5a0a5be17de2734a79e5b6a415412d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427264
Testability-Review: Kevin Wells <kevinwells@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>