[fdt] Always unzip the prebuilt to run
Before we were opening an already unzipped prebuilt binary if it
existed. The prebuilt system is only updating the .zip file so an
already unzipped binary might be stale so this is not a safe check.
Instead we always do the unzip step so that we actually pick up updates.

Change-Id: Iecb4089b7686d4e65974f1bb3bfd8ade6f4aabe3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428317
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Jayson Adams <shrike@google.com>
Reviewed-by: Alex Bogart <abbogart@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>