[wlan][sme] Make ConnectionLoss event accurate
Currently the ConnectionLoss event is triggered irrespective
of whether it was the AP or STA that initiated the disassoc/deauth.
This in turn results in inaccurate metrics.

This change ensures the event is triggered only when the disassoc/deauth
is triggered by the STA.

Bugs: 37823
Test: New tests added to wlan-sme-tests
Multiply: wlan-sme-tests

Change-Id: I493920a8ba4ced28f4522e7c41dd825401d1132c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428320
Reviewed-by: Kiet Tran <kiettran@google.com>
Testability-Review: Kiet Tran <kiettran@google.com>
Commit-Queue: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>