[cargo-gnaw] Refactor GN configration format
This is in preparation for the follow-on change that adds support for
generating executable GN targets for Cargo binary targets. Adding the
executable configuration was best done by first cleaning up how the
configuration for library targets is specified a bit. Assumptions that
"crates" == "packages" had crept in, which was natural since library
targets were the only targets supported previously and there's only one
of those per package.

The Cargo.toml looks very similar, but some of the unneeded flexibility
in the config format was removed. Target config specified on a
[gn.package.package-name.version] table (which is just a single table
now, not an array of tables) is now always the platform-independent
config, and platform-dependent config is specified through a new
"platform_cfg" table under that.

Also, previously the "foo" in "gn.crate.foo.version" was actually used
as the library target name; now it is the package name. For most
packages these names are the same.

Also, //third_party/rust_crates/Cargo.toml is cleaned up to re-jigger
configuration chunks into their proper sections, and cargo-gnaw's
README.md was updated and cleaned up a bit.

Bug: 59106
Change-Id: I6256793a46271ab1e21b4be28502e5ec8dd4f83a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/426195
Commit-Queue: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>