[fxtest] Exit `fx test` if build fails or cannot connect to device
When I'm testing a stack of CLs, I'll frequently run this loop:

```
% while fx test ... && git rebase --continue; do sleep 1; done
```

However, it turns out `fx test ...` only exits with a non-zero status if
the test fails, but exits with 0 if the build fails, or the test cannot
talk to the device. This patch changes `fx test` to fail in that case.

Test: Added test to make sure `fx test` will exit with a non-zero status
if the build fails or no running package server.

Change-Id: Ieaa0eeed2401dbe5ba9a17e307f1cc70f99a89f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427627
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Wesley Moy <wesleymoy@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Testability-Review: Kevin Wells <kevinwells@google.com>