[wlan] Cleanup prototypes of wlan final methods
Tricium has a linter that complains if 'override final' appears in a
class method prototype since it's redundant. I encountered this linter
in a previous commit to wlan::Device and am fixing the rest of them in
wlan with this change.

Test: None necessary. Pure syntactic change.

Change-Id: Ia46025fb41f68a3fc1e559ba79f5ca565c773a79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428443
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Eric Wang <eyw@google.com>
Testability-Review: Charles Celerier <chcl@google.com>