[crash_reports] Add a SnapshotManager
Many reports end up containing duplicate snapshots  because snapshot
requests are pooled. This is inefficeint when pool sizes are
large. To reduce the effect of large pools, reports will contain a
reference to a snapshot and the snapshot is shared between all reports
that are filed within a window of 5 seconds of one another.

One of the side effects of the previous implmentation of pooling was
that requests for a snapshot could timeout even after the snapshot had
already been collected if sycnhronous uploading took sufficiently long.
This implmentation aims to avoid that by not relying on fit::bridges to
advance the promise chain in the event of a timeout and instead checks
whether a snapshot has been collected before determining whether a
timeout has occurred.

Bug: 59514

TESTED=`fx test crash-reports-tests`

Change-Id: Ib89a5eb6cdab7373e71dad5ebf0a57588c9f5da0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427617
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>