[time] Wait for the first request before polling for time
This is a temporary hack done instead of checking for the network to
become available. Timekeeper doesn't start making requests for time
samples until it detects that network is available. This uses the first
fidl request as a signal that the network is ready. This is done
because the network check logic is currently unstable. Once it is
stable the logic should be moved to time source and this hack removed.

Testing: fx test httpsdate-time-source-tests

Change-Id: I7c4cea21e0d6408e4710b2c7654a70ba32d7679b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/427354
Reviewed-by: Jody Sankey <jsankey@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>