[fvm] Add fvm::Header getters.
Address off-by-one error for the maximum allocatable slices. Previously
fvm::FormatInfo::GetMaxAllocatableSlices() did not take into account
that there is an unused 0th entry in the allocation table.

Adds some helper functions to the fvm::Header structure to compute
various sizes and offsets from the header data.

Currently we have such getters spread around as toplevel functions that
accept the necessary parameters for the computation, and as members on
the fvm::FormatInfo structure.

The toplevel functions are inconvenient and error-prone as they require
the caller to separate out the correct data. The FormatInfo are better
but can't be used in the FVM driver itself because it maintains the
Header information in a large allocated block of the metadata which
can get out-of-sync with the separate header stored in FormatInfo.

As a result, I plan on migrating the existing uses to the new helpers on
the Header structure directly. This has the additional benefit of
bringing more consistent naming.

Currently the new getters are unused but there are some strategic
asserts that validate the getters produce the same answers as the
previous getters. I plan to let this bake for a bit to makes sure
everything matches, then convert stuff over.

Note the new definition of Header::GetAllocationTableUsedByteSize() +
Header::GetMetadataUsedBytes() and the current version of
FormatInfo::metadata_size() have different behavior in error cases.
metadata_size() currently returns the size the metadata *should* have
based on the current partition size. But Header::GetMetadataUsedBytes()
returns the size based on the pslice_count. I think when there's a
disagreement, it's safer to use pslice_count rather than infer the
portion of the table that's valid based on the disk size.

Change-Id: I277c0ef58c1a2b8fa951ee3445dfe6bdbd3ac3c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428418
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>