[io][conformance] Simplifies test setup with reusable FIDL setup method
Replaces overly-specific FIDL methods like GetDirectoryWithEmptyFile
with a general-purpose GetDirectory method. This method lets you provide
the structure of the entire filesystem, and will initialise it to that
state. The previous scattering of methods were each specific to a
particular test case, but this new method is reusable for any test
layout we want.

Will extend in a follow-up CL to allow adding VMO files and remote
directories to the initial state as well.

Change-Id: I6f7b7d90a21ca886331cd5621d644e278e61d74b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428601
Commit-Queue: Cameron Silvestrini <csilvestrini@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>