[archivist] GC stale ComponentLogStats
Cobalt often fails to receive accurate ComponentLogStats when a
component crashes restarts because when it disconnects the last clone of
the Arc<Mutex<ComponentLogStats>> gets dropped and a new object is
lazily instantiated if/when the component restarts.

This change introduces a timer that reviews Arc<ComponentLogStats> for
freshness and drops references to stale ones. This ensures a lower bound
on how long a Arc<ComponentLogStats> is retained, regardless of whether
the number of clients retaining them drops to 0 (as in the case with a
component crash or restart). The timer is currently set to trigger every
hour. For reference, currently Cobalt collects these metrics every 15
minutes.

BUG=56527
TEST=fx test //src/diagnostics/archivist:archivist_tests -- logs::stats::tests::component_stats_retained_then_dropped

Change-Id: Iee17169e91fc3958297c7e8d8f032e8216feec30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/428414
Reviewed-by: Fady Samuel <fsamuel@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Saman Sami <samans@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>