[fuzz] Resolve SSH credentials from manifest
The location of the default SSH credentials used in connecting to a
Fuchsia device has changed. This change updates fuzz to use a manifest
to find the appropriate path to the default SSH private key (identity)
and authorized_keys.

The aforementioned manifest is generated by a jiri hook on update and is
not expected to be absent in any Fuchsia tree unless the user has
removed it.

Change-Id: I8390e5fc90c1b449d161410235f287dfdba5aea5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419481
Testability-Review: Renato Mangini Dias <mangini@google.com>
Commit-Queue: Wesley Moy <wesleymoy@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>