[build] Restore zircon toolchain commands equivalence.
A recent change [1] moved the @ symbol used for response files
in quotes, but forgot to update the zircon_toolchain.gni file
as well. As a consequence, the build was not broken, but the
compare_toolchains.py script now complains that the commands
are not equivalent anymore between the Zircon and the Fuchsia
build.

This CL trivially fixes the issue.

Bug: 60296
Testing: The command [2] will succeed

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/429174
[2] zircon/public/canaries/compare_toolchains.py --clean /tmp/results

Change-Id: I3d460da2fedc85fc2a14e2cc6773686064e96302
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429614
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>