[build] Allow loadable_module() with PIC friendly toolchains.
The previous implementation of loadable_module() in the Fuchsia
build set _variant_shared to true unconditionally. This resulted
in every target of that type to always be built with an -shlib
companion toolchain.

Unfortunately, this doesn't work when trying to build the Zircon
userboot module, which will use its own PIC-friendly toolchain
suite (//zircon/kernel/lib/userabi/userboot:userboot_$cpu) that
does not have a -shared companion toolchain.

This CL fixes the issue by:

  1) Adding the `is_pic_default` parameter to variant_toolchain_suite(),
     implement it properly and document its use properly. Before this CL,
     the same feature was enabled by setting `shlib = true` in the call
     to zircon_toolchain_suite(), but its use was slightly obscure.

  2) Replace `shlib = true`  with `is_pic_default = true` in two
     Zircon-specific toolchain suite definitions.

  3) Ensure that loadable_module() will not use a companion-shlib
     toolchain for such toolchain instances.

For testing, the zircon/public/canaries/compare_toolchains.py script
was run to verify that this doesn't modify the commands of the
zircon-specific toolchains in the Fuchsia build.

The content of out/default/ after "fx gen" has also been compared before
and after the CL was applied, to verify that their content is identical.

Bug: 60360
Testing: Should not change the build. See above.

Change-Id: I7c350e6da4af6968945b4b07622761ce90aa101e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429615
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>