[utc][time][cm] Enable UTC clock propagation
TL;DR: If there are any issues with UTC time, revert this change.

This change enables the propagation of a UTC kernel clock owned by
component_manager and kept in sync by timekeeper. This clock gets
propagated through ELF Runner to all components. Appmgr, devhost,
fshost, and sshd are all configured to propagate the UTC clock if
it is set.

If there are any issues with UTC time being out of sync, this is likely
the culprit, as processes spawned by a process that is unaware of
userspace UTC kernel clocks will not receive the same UTC clock as the
rest of the system.

This is expected to be a transient issue only, as UTC time within the
kernel will be removed shortly.

Test: Manually boot product and verify that the time is current.
Test: fx iquery show timekeeper.cmx has a small delta between utc_nanos
      and utc_kernel_clock_value_nanos.
Test: fx test component-v1-time-unittests
Test: fx test component-v2-time-unittests

Change-Id: I259204998f4bd528e8a19cfa8b95caaf2fcfe939
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400971
Testability-Review: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>