[fidl][llcpp] Prevent overflowing unknown handle buffer
This asserts that we don't overflow the unknown handle buffer. A full
solution would close handles that overflow the buffer, but this adds
complexity - particularly so because there are different handle closing
implementations for handles and handle info.

Since this is a case that shouldn't occur ever currently, I favored
keeping the implementation simple and checking that it doesn't happen
with a ZX_ASSERT.

Fixed: 59999
Test: There are no tests because it will crash when this condition is
met and it is difficult to construct a test that can detect this
condition.

Change-Id: I3c6b48d39dd33102239dafa332844b82b6fbc0b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/429636
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>